solution/requirements.txt
fastapi[all]~=0.115.6
uvicorn~=0.34.0
asyncpg~=0.30.0
redis~=5.2.1
pydantic~=2.10.5
httpx~=0.28.1
argon2-cffi
PyJWT
cachetools
pycountry
uvloop
httptools
httpcore~=1.0.7
ujson~=5.10.0

dnspython~=2.7.0
idna~=3.10
anyio~=4.8.0
h11~=0.14.0
typing_extensions~=4.12.2
rich~=13.9.4
Pygments~=2.19.1
markdown-it-py~=3.0.0
PyYAML~=6.0.2
sniffio~=1.3.1
click~=8.1.8
mdurl~=0.1.2
typer~=0.15.1
shellingham~=1.5.4
MarkupSafe~=3.0.2
orjson~=3.10.14
certifi~=2024.12.14
starlette~=0.41.3
email_validator~=2.2.0
pydantic_core~=2.27.2
websockets~=14.1
httptools~=0.6.4
watchfiles~=1.0.4
greenlet~=3.1.1
annotated-types~=0.7.0
itsdangerous~=2.2.0
python-multipart~=0.0.20
Jinja2~=3.1.5
python-dotenv~=1.0.1
pydantic-settings~=2.7.1